Refactering code, creating person separete file
