+++ app.rb
@@ -1,6 +1,12 @@
+require_relative 'create_person'
+require_relative 'create_rental'
+require_relative 'create_book'
 require_relative 'book'
 require_relative 'student'
 require_relative 'teacher'
+require_relative 'list_book'
+require_relative 'list_people'
+require_relative 'list_rentals'
 class App
   attr_reader :people
 
@@ -49,113 +55,5 @@ class App
     answer = gets.chomp.to_i
     exit if answer == 7
     option_checker(answer)
-  end
-
-  def create_person
-    puts 'Create a student (1) or a teacher (2)? [Input the number]: '
-    answer = gets.chomp.to_i
-
-    case answer
-    when 1
-      create_student
-    when 2
-      create_teacher
-    end
-
-    run
-  end
-
-  def create_student
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Name: '
-    name = gets.chomp
-
-    print 'Has parent permission? [Y/N]'
-    permission = gets.chomp.upcase
-    case permission
-    when 'Y'
-      permission = true
-    when 'N'
-      permission = false
-    end
-
-    student = Student.new(age, nil, name, parent_permission: permission)
-    @people.push(student)
-    puts 'A student created successfully'
-  end
-
-  def create_teacher
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Specialization: '
-    specialization = gets.chomp
-    print 'Name: '
-    name = gets.chomp
-
-    teacher = Teacher.new(age, specialization, name)
-    @people.push(teacher)
-    puts 'Teacher created successfully'
-  end
-
-  def create_book
-    print 'Title: '
-    title = gets.chomp
-    print 'Author: '
-    author = gets.chomp
-    book = Book.new(title, author)
-    @books.push(book)
-    puts 'Book created successfully'
-    run
-  end
-
-  def create_rental
-    puts 'Select the book from the following list by number (not ID)'
-    @books.each_with_index do |book, idx|
-      puts "#{idx}) Title: '#{book.title}', Author: '#{book.author}'"
-    end
-    selected_id_book = gets.chomp.to_i
-    print 'Date: '
-    date = gets.chomp
-
-    puts 'Select a person from the following list by number (not ID)'
-    @people.each_with_index do |person, idx|
-      puts "#{idx}) [#{person.class.name}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-    selected_id_person = gets.chomp.to_i
-
-    rental = Rental.new(date, @books[selected_id_book], @people[selected_id_person])
-    @rentals.push(rental)
-    puts 'Rental created successfully'
-    run
-  end
-
-  def list_books
-    puts ''
-    puts 'List of all the books in the library:'
-    puts ''
-    @books.each do |book|
-      puts "Title: #{book.title}"
-      puts "Author: #{book.author}"
-    end
-    run
-  end
-
-  def list_people
-    puts ''
-    puts 'List of all people:'
-    @people.each do |person|
-      puts "[#{person.class.name}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-    run
-  end
-
-  def list_rentals
-    print 'ID of person; '
-    id = gets.chomp.to_i
-    rentals = @rentals.select { |rental| id == rental.person.id }
-
-    rentals.each { |item| puts "Date: #{item.date},  Book #{item.book.title}, by #{item.book.author}" }
-    run
   end
 end
