+++ decorate_pattern/capitalize_decorator.rb
