+++ decorate_pattern/decorator.rb
