+++ decorate_pattern/nameable.rb
