+++ decorate_pattern/trimmer_decorator.rb
